feat: 교육콘텐츠 삭제 exception 추가

+++ src/curriculum/modules/educational-content/educational-content.resolver.ts
@@ -96,7 +96,15 @@ export class EducationalContentResolver {
     @Args('input') input: DeleteEducationalContentInput,
   ): Promise<boolean> {
     const { id } = input;
+
+    const educationalContent =
+      await this.educationalContentService.getEducationalContent({ id });
+    if (educationalContent === null) {
+      throw new NotFoundException('educationContent not found');
+    }
+
     await this.educationalContentService.deleteEducationalContent(id);
+
     return true;
   }
